Rename colossusx package declarations
