+++ consensus/colossusx/sealer.go
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the cypherBFT library. If not, see <http://www.gnu.org/licenses/>.
 
-package ethash
+package colossusx
 
 import (
 	crand "crypto/rand"
